Split processor
