+++ packages/contracts/source/contracts/processor/block-processor-result.ts
@@ -1,0 +1,7 @@
+import { TransactionReceipt } from "../evm/index.js";
+
+export interface BlockProcessorResult {
+	success: boolean;
+	receipts: Map<string, TransactionReceipt>;
+	gasUsed: number;
+}
