+++ packages/contracts/source/contracts/processor/handler.ts
@@ -1,0 +1,5 @@
+import { ProcessableUnit } from "./processable-unit.js";
+
+export interface Handler {
+	execute(unit: ProcessableUnit): Promise<void>;
+}
