+++ packages/contracts/source/contracts/processor/verifier.ts
@@ -1,0 +1,5 @@
+import { ProcessableUnit } from "./processable-unit.js";
+
+export interface Verifier {
+	verify(unit: ProcessableUnit): Promise<void>;
+}
